Adds react-redux-starter-kit to examples
This boilerplate is still a work in progress but is currently functional, it could just use some improvement. I'm actively working on it and am focused on keeping it up to date with the latest Redux and React Router API changes so people can get an idea of how to use both.