fix(real-world-example): lock react-router version
Previously the real world example was not working due to a difference in
the file structure in react-router 1.0.0-beta4

The following error was being raised:

 > Module not found: Error: Cannot resolve module
 > 'react-router/lib/BrowserHistory' in /tmp/redux/examples/real-world

The version has been locked to react-router 1.0.0-beta3 to ensure the
example still runs. When the documentation has been updated for
react-router it would be ideal to update this to use the latest version
so the demo is up to date.