Added support to combineReducers to make the coding of loading state easier.
With the existing implmentation of combineReducers if you try to add an action to load
the state every reducer will get the entire state passed to them and would need to have
knowledge of where they are in the state tree to lookup their new state.

I've added a special HYDRATE action type (@@redux/HYDRATE) which combineReducers looks for and
handles as a special case passing down only the state for each reducer.

This means that each reducer can opt in to the load and choose how it wants to handle it.
Usually this will just be returning action.state.