Correct reference for recipe documentation.
This is more a suggestion for more general fixes to recipe docs links.  I'm relatively new to redux and was trying to follow this server rendering recipe but saw incorrect files (according to current master branch) referenced when trying to complete recipe steps.  Looking back in the project history, it looks like v3.0.6 may be the last tag that is compatible with the suggested changes. 

Maybe someone who knows the project better can confirm that this is the correct tag to reference for these examples?

I assume this is preferred over rewriting all of the recipes right now.  I just want accurate docs. What do others think?