Remove `bindActionsCreators()` warning
This commit makes so that there is no warning printed
out when calling `bindActionCreators()` function with
the first argument containing key "default" (which is
not a function).

This is done in order to allow library users to
store their actions in the module in which
actions _also_ exported as a `default` object.
So, commit just skips that object.