Fix broken Promise in thunk action creator
The current `onRejected` function resolves with `undefined` because it
returns the `console.log()` expression. This means that if the `fetch()`
errors, the `onRejected` in the first `then()` will resolve with
`undefined`, causing `json` in the second `then()` to be `undefined`.
Calling `receivePosts()` with `json === undefined` will error on
`json.data`.

There's no reason to use two `then()` calls. But if two `then()` calls
are desired, the second `then()` must happen before the `onRejected`
function, because the `onRejected` function should resolve with
something so it doesn't have an unhandled Promise rejection. In the fix,
the `onRejected` function resolves with the `error`. To have the second
`then()`, you could nest the `then()` inside the `onResolved` function
like so:
  .then(
    response => Promise.resolve(response.json)
      .then(
        json => dispatch(receivePosts(subreddit, json))
      )
  )
But there's no point in doing that.