Convert postId param to number before using
Going through this tutorial in the docs, I found that match.params.postId is a string, whereas the postId on the post object is a number (line 252).
Propose wrapping this in the Number() function to avoid state.posts.find(post => post.id === postId) returning undefined. 

e.g 2 === '2' returns false in line 255.