fix: codesandbox embeds mobile scrolling
Currently the embeds for codesandbox.io do not allow you to scroll on
mobile devices (tested on Samsung Galaxy S22 Ultra). This is a problem
because users are unable to view the whole project as they are reading
through the amazing documentation and examples that have been created.
Fix this by updating the src url to include the codemirror option which
uses the CodeMirror editor instead of Monaco and also reduces the embed
size as well