Remove Google Analytics
Closes #3815 

I enabled Netlify Analytics and these seem to cover our needs from the perspective of knowing which pages are most popular. Even better, they're retroactive and already have data collected. Super bonus: They're server-side and don't require a tracking script on the page (which can be blocked and is a privacy concern).